Move crypto code to a new class
