+++ Phase2/src/UserList.java
@@ -97,18 +97,7 @@ public class UserList implements java.io.Serializable {
         else
             return false;
     }
-    
-    static byte[] getHash(String input)
-    {
-        byte[] toHash = input.getBytes();
-        // TODO: other algorithm?
-        try {
-            MessageDigest digest = MessageDigest.getInstance("SHA1", "BC");
-            digest.update(toHash);
-            return digest.digest();
-        }
-        catch(Exception e) { e.printStackTrace(System.err); return null; }
-    }
+
     static boolean comparePasswords(String password, User UserInfo)
     {
             return Arrays.equals(getHash(password), UserInfo.getPasswordHash());
@@ -240,17 +229,5 @@ class User implements java.io.Serializable {
             }
         }
         return false;
-    }
-    
-     static byte[] getHash(String input)
-    {
-        byte[] toHash = input.getBytes();
-        // TODO: other algorithm?
-        try {
-            MessageDigest digest = MessageDigest.getInstance("SHA1", "BC");
-            digest.update(toHash);
-            return digest.digest();
-        }
-        catch(Exception e) { e.printStackTrace(System.err); return null; }
     }
 }
